+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-### Added 
+### Added
 
 - `attributes` field in `/list-resources` and `/list-subjects` endpoint
 - Sanitize service support
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- The source-url transfer method now works with File Scan.
+- The source-url transfer method now works with File Scan and Sanitize.
 
 ### Removed